requirements.txt
Django==4.2.7
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
django-filter==23.3
django-summernote==0.8.20.0
Pillow==10.1.0